init: Initial implementation.
Running "bob init <dir>" will populate a directory with a configuration
file and a set of scripts that will be used by the build.  This allows
easy access to edit them, have separate build configurations, etc.  For
now no script or hooks to run them but that will come shortly.